create KarhooMailMetaInfoComposer
