+++ KarhooUISDK/CommonUI/MailComposer/KarhooFeedbackEmailComposer.swift
@@ -22,6 +22,8 @@ public final class KarhooFeedbackEmailComposer: NSObject, FeedbackEmailComposer 
     private weak var viewController: UIViewController?
 
     private let mailComposer = MailComposer()
+    
+    private var mailMetaInfoComposer: KarhooMailMetaInfoComposerProtocol = KarhooMailMetaInfoComposer()
 
     public func set(parent: UIViewController) {
         self.viewController = parent
@@ -35,7 +37,7 @@ public final class KarhooFeedbackEmailComposer: NSObject, FeedbackEmailComposer 
         _ = mailComposer.presentMailController(from: viewController!,
                                                subject: UITexts.SupportMailMessage.feedbackEmailSubject,
                                                recipients: [UITexts.SupportMailMessage.feedbackEmailAddress],
-                                               body: mailMetaInfo())
+                                               body: mailMetaInfoComposer.mailMetaInfo())
         return true
     }
 
@@ -65,20 +67,7 @@ public final class KarhooFeedbackEmailComposer: NSObject, FeedbackEmailComposer 
 
     // MARK: private
 
-    private func mailMetaInfo() -> String {
-        let info = """
-                   \n \n \n
-                   \(UITexts.SupportMailMessage.feedbackMailMessage)
-                   ------------------
-                   Application: \(appName()) \(appVersion())
-                   Device: \(platform())
-                   System: \(systemName()) \(systemVersion())
-                   Locale: \(currentLocale())
-                   User: \(userInfo())
-                   ------------------
-                   """
-        return info
-    }
+    
 
     private func tripReportBody(trip: TripInfo) -> String {
         let info = """
@@ -86,65 +75,8 @@ public final class KarhooFeedbackEmailComposer: NSObject, FeedbackEmailComposer 
                    ------------------
                    Trip: \(trip.displayId)
                    ------------------
-                   \(mailMetaInfo())
+                   \(mailMetaInfoComposer.mailMetaInfo())
                    """
         return info
-    }
-
-    private func userInfo() -> String {
-        guard let user = Karhoo.getUserService().getCurrentUser() else {
-            return ""
-        }
-
-        return """
-               \(UITexts.User.email): \(user.email)
-               \(UITexts.User.mobilePhone): \(user.mobileNumber)
-               \(UITexts.User.firstName): \(user.firstName)
-               \(UITexts.User.lastName): \(user.lastName)
-               """
-    }
-}
-
-private extension KarhooFeedbackEmailComposer {
-    func appName() -> String {
-        guard let dictionary = Bundle.main.infoDictionary else {
-                    return ""
-        }
-        if let version: String = dictionary["CFBundleDisplayName"] as? String {
-            return version
-        } else {
-            return ""
-        }
-    }
-    
-    func appVersion() -> String {
-        guard let dictionary = Bundle.main.infoDictionary else {
-                    return ""
-        }
-        if let version: String = dictionary["CFBundleShortVersionString"] as? String {
-            return version
-        } else {
-            return ""
-        }
-    }
-
-    func systemName() -> String {
-        return UIDevice.current.systemName
-    }
-
-    func systemVersion() -> String {
-        return UIDevice.current.systemVersion
-    }
-
-func currentLocale() -> String {
-        return NSLocale.current.identifier
-    }
-
-    func platform() -> String {
-        var sysinfo = utsname()
-        uname(&sysinfo) // ignore return value
-        return String(bytes: Data(bytes: &sysinfo.machine,
-                                  count: Int(_SYS_NAMELEN)),
-                      encoding: .ascii)!.trimmingCharacters(in: .controlCharacters)
     }
 }
